requirements.txt
langchain-openai>=1.0.2
langchain-anthropic>=1.0.3
langchain_tavily>=0.2.13
pydantic>=2.0.0
rich>=14.0.0
jupyter>=1.0.0
ipykernel>=6.20.0
tavily-python>=0.5.0
httpx>=0.28.1
markdownify>=1.2.0
deepagents>=0.2.6
python-dotenv>=1.0.0
langgraph-cli[inmem]>=0.1.55
langchain-google-genai>=3.1.0
langchain-experimental>=0.0.62
